five

+++ exercises/conversions/using_as.rs
@@ -7,14 +7,19 @@
 // The goal is to make sure that the division does not fail to compile and
 // returns the proper type.
 //
+// 在Rust中，类型转换是通过使用`as`操作符来完成的。请注意，as `运算符不仅在类型转换时使用。
+// 它还有助于重命名导入。
+//
+// 我们的目标是确保除法不会编译失败，并返回正确的类型。
+//
 // Execute `rustlings hint using_as` or use the `hint` watch subcommand for a
 // hint.
 
-// I AM NOT DONE
+
 
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
-    total / values.len()
+    total / values.len() as f64
 }
 
 fn main() {
